Move OAuth2 patching to an initializer
